% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVar.R
\name{combined_likelihood}
\alias{combined_likelihood}
\title{Posterior Mean Estimator of the Variance with Integrated Combined Likelihood}
\usage{
combined_likelihood(pheno_long, N, ns)
}
\arguments{
\item{G}{A s by p matrix of genotypes, where s is the number of strains and p is the number of snps to be tested. This matrix can have missing values, and each SNP should be coded as 0, 1, 0.5 or NA}

\item{y}{A N length vector of phenotypes for each individual organism, where N is the total number of individuals}

\item{strains}{A s by N incidence matrix that maps every individual to a strain}

\item{X}{A s by q matrix of covariates (optional)}

\item{K}{A s by s genomic relationship matrix. Will be calculated if unspecified.}

\item{weights}{A string specifying the weights to be used. The following are permitted: "none", "samplevars","EB", "limma", "counts", and "user"}

\item{user_weights}{A s length vector of weights for each strain, used if weights = "user"}
}
\value{
A list containing:
\itemize{
\item{pvalue: A p length vector of p-values for every snp}
\item{ML0: A p length vector of log maximum likelihood under the null hypothesis}
\item{ML1: A p length vector of log maximum likelihood under the alternative hypothesis}
\item{beta: A p length vector of regression parameter estimate for slope under the alternative hypothesis}
\item{s2: A p length vector of the sum of variance components under the null hypothesis}
\item{h2: A p length vector of the heritability estimate under the null hypothesis}
}
}
\description{
Performs a genome scan on heteroscedastic data.
}
